Corrections to `MapView` after the introduction of `ValueOrBytes`. (#3046)
## Motivation

Several additional suggestions were forgotten in the PR
https://github.com/linera-io/linera-protocol/pull/3036

## Proposal

The following corrections were made:
* The `ValueOrBytes` is marked as private.
* The `count` used for `SetView`, `CollectionView`,
`ReentrantCollectionView`, `MapView` and `KeyvalueStoreView` are `usize`
which makes them architecture dependent. We switch to u32.
* The `FnMut(I, V)` forces a clone which is not always needed. So, it
was replaced by `FnMut(I, Cow<'a, V>)`. Actually, the clone still
occurs, but if someone were to want to just access the reference, it
would be possible.

Unfortunately, the use of the Cow complexifies the call to the
`for_each_key_value`.

## Test Plan

The CI.

## Release Plan

This is a breaking change with TestNEt / DevNet as it changes the hashes
values.

## Links

None.