Make the chain manager a `View`. (#3135)
## Motivation

Currently the chain manager is serialized as a whole and stored in a
`RegisterView` in the chain state view. Since it contains blobs it can
be very large, and the blobs are not needed every time the chain manager
is loaded.

## Proposal

Make the chain manager a `View`, and put the blobs in a `MapView`.

## Test Plan

This doesn't change any logic, so CI should catch regressions. (In fact,
it already did: https://github.com/linera-io/linera-protocol/pull/3133)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- In preparation for:
https://github.com/linera-io/linera-protocol/issues/3048
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)