Improve the tests to the Views. (#3166)
## Motivation

Recently, a long-standing bug was found in the `RegisterView`. Thus it makes sense to strenghten the tests.

## Proposal

Some implementing of `TestView` were done for `SetView` / `QueueView`.

The test that detects the found problem for the `RegisterView` was
extended to all the view supporting `TestView`.

## Test Plan

Simply improve the tests.

## Release Plan

Nothing relevant.