Simplify chain manager logic. (#3178)
## Motivation

We are using the chain manager's `locked` field for two purposes:
* In validators, this is meant to track _this validator's_ locked block,
i.e. the latest proposal it signed to confirm.
* In the client, it tracks the highest validated block (or fast
proposal) we have seen, i.e. the highest block that _might_ be locked
for any honest validator, so that the client knows which block it should
re-propose in the next round.

This is confusing, and makes it more difficult to tackle issues like
https://github.com/linera-io/linera-protocol/issues/2971.

## Proposal

Use the `locked` field for the first purpose, both in clients and
validators: to track the highest block that _might_ be locked for any of
the validators.

In a validator's chain manager, the block that is locked for _that
validator_ is the one it has last voted to confirm, i.e. the
`confirmed_vote` field.

## Test Plan

The tests already cover lots of scenarios involving locked blocks, and
should catch any regressions. They have been updated where appropriate.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes https://github.com/linera-io/linera-protocol/issues/3170.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)