Update the storage service reading. (#3011)
## Motivation

The `read_entries` code of the storage-service is doing a blocking loop.
That is inefficient from the viewpoint of concurrency.

## Proposal

The following was done:
* A vector of handles is created.
* The handles are processed and then merged.
* On the server side, the code is changed so that it is no longer
assumed that the request arrive in order.

## Test Plan

The CI proved adequate. The `test_storage_service_big_raw_write` did
catch the problem with the ordering of the requests.

## Release Plan

Not relevant as we do not use so far the `storage-service` in TestNet /
DevNet.