Use `Mutex` inside `ServiceRuntime` to make it `Send` (#3201)
## Motivation

<!--
Briefly describe the goal(s) of this PR.
-->
Using the `ServiceRuntime` in services that use GraphQL requires
wrapping the `ServiceRuntime` inside an `Arc<Mutex<_>>`, because the
`async_graphql` crate requires the handler to be `Send`. This is bit
burdensome for application developers, and is redundant because the
`Service` is single-threaded and uses `Cell` for interior mutability.

## Proposal

<!--
Summarize the proposed changes and how they address the goal(s) stated
above.
-->
Replace usages of `Cell` with `Mutex` inside the `ServiceRuntime`, and
also do the same for `MockServiceRuntime`.

Update the examples to remove the redundant `Mutex`es.

## Test Plan

<!--
Explain how you made sure that the changes are correct and that they
perform as intended.

Please describe testing protocols (CI, manual tests, benchmarks, etc) in
a way that others
can reproduce the results.
-->
This is an internal refactor, and application end-to-end tests should
catch any regressions.

## Release Plan

<!--
If this PR targets the `main` branch, **keep the applicable lines** to
indicate if you
recommend the changes to be picked in release branches, SDKs, and
hotfixes.

This generally concerns only bug fixes.

Note that altering the public protocol (e.g. transaction format, WASM
syscalls) or storage
formats requires a new deployment.
-->
- Nothing to do / These changes follow the usual release cycle, because
it only affects the SDK.

## Links

<!--
Optional section for related PRs, related issues, and other references.

If needed, please create issues to track future improvements and link
them here.
-->
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)