Change mutations to schedule operations using a system API (#3217)
## Motivation

Previously, services were required to use GraphQL in order to support
creating blocks with operations. The design was that each mutation query
would return a serialized operation as a vector of bytes. This is a
constraint that's not really needed, and requires extra work by
application developers.

## Proposal

Add a `schedule_operation` system API, which adds an operation to a list
of scheduled operations. More than one operation can be scheduled to be
added to the proposed block. The list is then returned back to the
client (in the node service), which then uses it to build a block
proposal and send it to the validators.

The SDK was updated to add `ServiceRuntime::schedule_raw_operation` to
send the operation as a vector of bytes, and also add a
`ServiceRuntime::schedule_operation` method which accepts any
`Serialize` type and handles serialization for the application.

All examples were updated to use the new API, either manually if the
mutation object is implemented manually or by the changes to the
`GraphQLMutationRoot` derive macro. Unfortunately, `async_graphql`
requires the methods to return _something_ that implements `OutputType`,
so for now they return an empty array of bytes.

## Test Plan

CI should catch any regressions by the switch to use the new API.

## Release Plan

- Backporting is not possible but we may want to deploy a new `devnet`
and release a new SDK soon.

In theory only the client is affected, but since the API is available to
services, when running a contract the validators could query a service
that runs as an oracle and the behavior would differ if this change is
included in the validator or not. In both cases an error will be
returned and the block rejected, but the errors would be different. To
mitigate risk, this should be released together with validator upgrades.