Several corrections to `linera-execution` code. (#3231)
## Motivation

The work on the integration of the EVM into the `linera-execution` led
to code reading and so to identification of some ways to improve the
code.

## Proposal

The following was done:
* The functions `user` and `user_without_abi` are only used for testing
purposes and so marked as such.
* The functions `find_dependencies` and
`describe_applications_with_dependencies` are always used with an empty
last argument in the code. The code is thus simplified.
* Many `clone()` operations in `system.rs` are not needed even though
clippy did not detect this.
* The use of `Sender` / `oneshot::Sender` is inconcistent in the code.

## Test Plan

The CI should detect it.

## Release Plan

Follow the normal release plan.

## Links

None.