Enable the listing of `BlobIds` from the command line. (#3161)
## Motivation

Knowing the blobs that are in storage is useful for debugging. We
implement this feature here.


## Proposal

The functionality is implemented from the `db_storage`. The problem is
that the implementation depends on the serialization so a test is added
for this.

## Test Plan

The CI.

For the feature itself, it was tested locally from the command line for
the non-fungible test.

## Release Plan

A new functionality is added, but no break of storage schema occurs.

## Links

None.