Implement `list_root_keys` to the `AdminKeyValueStore` (#3142)
## Motivation

We need to have the feature of accessing to the list of root keys in
order to have mutation function and accessing to the list of chain ids
of a storage.

Fixes #3085 

## Proposal

The implementation causes some problems:
* For ScyllaDb the implementation is very easy since the root_key is
used as a partition key.
* For DynamoDb we can implement the root_key as a partition key, but
DynamoDb forbids iterating and determining all the partition keys. So,
we need to keep track of the keys.
* For RocksDb / StorageService / IndexedDb we need to keep track of the
list of root_keys.
* For RocksDb, this led to a simplification since the edge case of
having a key of the form `[255, ..., 255]` disappears. This corrects a
problem and a test is added to detect it.
* The result of the `list_root_keys` will not be the same on different
storage. If storage has been created with `fn create`, some keys were
written but later deleted then in DynamoDb, RocksDb, storage-service,
IndexedDb the root key will show up as existing while in ScyllaDb, the
root key will not be visible.

The writing of the root key occurs when a `write_batch` is done.

## Test Plan

One test has been added for this feature.

## Release Plan

No impact on the TestNet / DevNet. It can follow the normal release
plan.

## Links

None.