Fix dangling endpoint to chain worker actor that failed to start (#3298)
Testnet is currently experiencing some panics when an attempt is made to
send a request to a chain worker actor. The panic message says that the
endpoint is closed. One possible cause is #3295, where a storage error
could cause an endpoint to be added to the chain worker cache but its
respective actor task wouldn't start.

Refactor the `ChainWorkerActor` so that it has a `run` method that loads
the chain state from storage and then handles the incoming requests. If
loading the state fails, then the error is reported to the next
requester, and loading will be reattempted while the actor is running
and receiving requests.

CI should catch any regressions caused by this refactor.

- These changes should be backported to the latest `devnet` branch, then
    - be released in a validator hotfix.
- These changes should be backported to the latest `testnet` branch,
then
    - be released in a validator hotfix.

This needs to be hotfixed because it may fix a bug that's in production.

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- Closes #3295