Introduce long running mode to linera benchmark (#3265)
## Motivation

We need a way of proposing blocks to a network at a fixed BPS rate

## Proposal

Introduce a long running mode that you can specify a BPS rate, and it'll attempt to reach that. This is the most naive version, doesn't scale very well.
We need to figure out how much we're gonna paralellize these before having another binary run this in different processes with different wallets.
My proposal is that we implement until Version 3, then instead of the "worker tasks" we split these into separate processes with separate wallets.

## Test Plan

Ran it locally against a local kind network.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.