Add cargo-spellcheck config and dictionary; fix typos. (#3331)
## Motivation

We are getting lots of PRs that fix individual typos. That is a bit
noisy, and wastes a lot of CI time.

## Proposal

Use `cargo-spellcheck` and fix lots of typos at once.

Add a `cargo-spellcheck` configuration and word list. Unfortunately one
has to locally copy this to the hunspell directory, or create a symbolic
link.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)