Gracefully shutdown benchmark (#3326)
## Motivation

Right now if you interrupt the benchmark, it might leave you in a state where the local node is inconsistent with the validators, the wallet hasn't been saved... etc

## Proposal

Gracefully shutdown the benchmark on interruption signals, making sure we never exit in the middle of sending a block proposal, and only saving the wallet on program exit.

## Test Plan

Ran the benchmark locally, did Ctrl+C and saw it gracefully exit.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.