Introduce --with-faucet for net up using the default chain (#3324)
## Motivation

Right now we have a `--with-faucet-chain` option that both enables starting a faucet and also specifies what chain the faucet will be running on.
`linera faucet` however has a slightly different behavior, in which if you don't specify the chain, it just uses the default chain.

## Proposal

Introduce a `--with-faucet` option that controls whether a faucet will be started, which will by default start a faucet on the default chain, unless the `--faucet-chain` argument is provided.

## Test Plan

Run it locally, see the faucet being started using the default chain as expected, and previous behavior also works

## Release Plan

- Nothing to do / These changes follow the usual release cycle.