Make sure we include end_value in generated Prometheus buckets (#3363)
Right now we're generating the buckets, but it's not in an inclusive way, and it should be. For latency metrics, for example, you have `end_value` being 500ms, but one of the buckets is 159. Since we're using a multiplication factor of 3, the next bucket would go over 500ms, so we don't include it...

Make sure we always include the `end_value` bucket

CI + start a network locally and see the correct buckets exported

- Nothing to do / These changes follow the usual release cycle.