Implement the functionalities of listing the `ChainIds`. (#3355)
## Motivation

We implement here the functionality of listing the Chain IDs from a storage.

## Proposal

The implementation follows the same path as from the implementation of
the listing of the BlobIds by using the `list_root_keys`.
The `list_root_keys` functionality is removed from the CLI.

## Test Plan

The functionality was tested locally by running the non-fungible test and listing the Chain IDs.

## Release Plan

Normal release schedule.