Use service to query ethereum node in Ethereum Tracker example (#3144)
## Motivation

While working on PR #2631, the end-to-end test for the Ethereum Tracker
example started failing. That happened because the updated HTTP request
API also included the response headers in the oracle recording, and a
"Date" header made the response non-deterministic.

## Proposal

Use the service to perform the HTTP request, and only send back to the
contract the actual deterministic information that matters.

In order to do so, the `EthereumClient` had to be split in two, so that
it could use either the contract system APIs or the service system APIs.

## Test Plan

The end-to-end test passes locally in my repository when this fix is
applied to the changes in PR #2631.

## Release Plan

- Nothing to do / These changes follow the usual release cycle, as it
only changes the example and the SDK.