Remove `mut` from `ServiceRuntime::http_request` (#3435)
## Motivation

PR #2631 updated the runtime APIs to perform HTTP requests, but
unfortunately the `ServiceRuntime::http_request` method had an
unnecessary `&mut self` receiver. This makes it different from all other
`ServiceRuntime` methods.

## Proposal

Change the receiver to be `&self`, and do the same in the
`MockServiceRuntime`.

## Test Plan

CI should catch any regressions by this refactor.

## Release Plan

- Nothing to do / These changes follow the usual release cycle, because
this fixes an issue with a feature that hasn't yet been released.