Make Evm a part of the supported virtual machines. (#3386)
## Motivation

We extend here the support for EVM. This means extending the
`BytecodeId` by adding the `VmRuntime`to it.

## Proposal

The following changes have been done:
* A type `VmRuntime` has been introduced that covers both `Wasm` (Native
linera) and `Evm`.
* The `VmRuntime` is added to the `BytecodeId`. This is a reasonable way
to make the VM a part of the type description.
* We cannot put the `VmRuntime` in the `UserApplicationDescription`
since this forces the `create_application` to have a `VmRuntime` type.
While innocent, this changes is not adequate since the
`create_application` is used in smart contract code and we want to avoid
this part to seep into the smart contract code.
* The `VmRuntime,` does not depend on any feature. This is because they
have to pass the wit barrier and the wit-types cannot depend on
features.
* The `VmRuntime` has been added to the `linera-base`. This is
unavoidable because `BytecodeId` depends on it.
* The `load_contract/load_service` is simplified into one single code
that panics when relevant features have not been installed.
* The `--vm-runtime` is added to the function calls like
`publish_application`. When missing it defaults to `VmRuntime::Wasm`.
* A typo in `contract_runtime_apis.rs` for `ApplicationDescription` is
corrected.

## Test Plan

The CI. No test has been added to test the functionality.

## Release Plan

Normal release plan.

## Links

None.