Add option to not close chains on benchmark (#3447)
## Motivation

Closing chains actually takes a surprisingly long time. The client will actually return fairly quickly, but the validators will be executing stuff for a while after the `CloseChain` operation is done. Also latency spikes while the validator is doing that.
It's probably something to be investigated as well, tbh.

## Proposal

For now, add an option to the benchmark to just not close the chains that are created for the benchmark (don't close them at the end), so that iteration with a lot of chains is faster.

## Test Plan

Run it with the option, see that we exit quicker and the validators are not doing a bunch of stuff after the benchmark is exited.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.