Query prometheus metrics from benchmark (#3473)
## Motivation

Right now, the only way of knowing if the benchmark is completely destroying validators is by checking Grafana for the metrics (which have a 2m delay), and stopping the benchmark manually. This is not ideal, as if latency becomes too high, the validators can take a very long time to recover after it.

## Proposal

Introduce a `--health_check_endpoints` to `linera benchmark`, that if provided, will query the specified metrics host:port pairs for metrics, to determine the health of the validators. For now this health is determined by the p99 proxy latency being below 400 ms.
We're querying the metrics endpoint from the proxy directly with this, so it works on all different kinds of deployments (docker compose, kubernetes, just binaries running locally, etc). The user is responsible for making sure the metrics endpoint is available at the provided endpoints (might need to port forward the metrics port, etc).
We'll query the endpoints once every 5 seconds, and calculate the p99 of the different validators in those last 5s. If we see one validator being above the 400 ms threshold, automatically stop the benchmark. We calculate the p99 doing the linear interpolation manually, based on the histogram data we get from the endpoints.

## Test Plan

Ran a network locally, ran the benchmark against it, saw the p99 values being printed matched what Grafana reported as well, and that if we cross the threshold the benchmark gets stopped automatically.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.