[benchmark] Don't obliterate validators as much when wrapping up (#3474)
## Motivation

Right now if we have the `--close-chains` option on, we'll close all the chains in parallel. If we don't, we'll process all the inboxes in parallel also before saving the benchmark chains to the wallet.
This is not ideal because depending on how many chains we're using, this can pound the validators pretty hard, and drive latency up to the point of us getting a bunch of timeouts when wrapping things up.
Also, when latency spikes up too much, the validators take a while to recover.

## Proposal

Introduce a `--wrap-up-max-in-flight` argument to limit the concurrency when wrapping things up

## Test Plan

Tested locally, no timeouts anymore, and things actually go faster (if we timeout, we retry, and this can take a while until we actually eventually succeed within the timeout constraint)

## Release Plan

- Nothing to do / These changes follow the usual release cycle.