Make Account* keys an enum of Ed25519 and Secp256k1. (#3465)
## Motivation

We want users to use their EVM wallets (like Metamask or others) when
interacting with Linera (which technically means allowing for secp256k1
keys to be used as chain owners).

## Proposal

Change `Account*` abstractions from a newtype struct to an enum with two
variants: Ed25519 and secp256k1.


## Test Plan

A crowdfunding example was modified to have two chains with different
key schemes for chains' owners. This unfortunately didn't excersise
signature verification as we don't do that for the Linera application
testing framework.

Benchmarks were also modified to generate chains with both key variants
and exersise what was not possible in the example.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.