Add exists chain to TestValidator to test new created chain (#3484)
## Motivation

If we create new chain with `open_chain` in `contract.rs`, it won't be
maintained by `TestValidator` in integration tests. Thus we won't be
able to test some functionalities of the new chain in integration tests.
For example, if we would like to call operation from another user chain
to the created chain's application, the user chain should register the
application firstly, then call it. At that time we'll get error `Chain
not found` with the `TestValidator`.

## Proposal

If we could add exists chain to `TestValidator`, then I think we'll be
able to test those created chains in integration tests.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

---------

Signed-off-by: Zhao KK <kikakkz@hotmail.com>
Signed-off-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>