Support custom `ResourceControlPolicy` in integration tests (#3505)
## Motivation

In integration tests, the `ResourceControlPolicy` was hardcoded to use
the default value. With the addition of an HTTP allow list (#3427), the
tests aren't able to test any HTTP requests.

## Proposal

Allow changing the `ResourceControlPolicy` used inside a test, allowing
to add hosts to the allow list. For the policy change to work, a new
committee and epoch must be published. Therefore the epoch and committee
can't be assumed to be constant anymore.

## Test Plan

Wrote some (upcoming) tests that try to perform an HTTP request without
changing the policy and after adding the host to the policy, and ensured
that the former fails while the latter succeeds.

## Release Plan

- These changes follow the usual release cycle, because this will help
with tests that use a feature that has not yet been released.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)