Simplify testing GraphGL mutations in integration tests (#3506)
## Motivation

It is useful in integration tests to test the flow from a service
mutation up to the contract execution. This could be done previously by
calling `graphql_query` and then manually adding a new block. However,
this had some limitations, as for example if the service called another
service which created an operation for the called application could not
easily be added.

Also, given how frequent GraphQL mutations are used in applications, it
makes sense to make testing it as seamless as possible.

## Proposal

Add an `ActiveChain::graphql_mutation` helper method, which queries the
service and attempts to create a block with the scheduled operations.

## Test Plan

Tested manually in some upcoming integration tests for HTTP requests.

## Release Plan

- These changes follow the usual release cycle, because they contain
non-breaking new features.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)