Implement the EVM test in the end-to-end test. (#3479)
## Motivation

The end-to-end test for Evm is needed as a tool.

## Proposal

The following is done:
* The compilation functionalities are moved to a `solidity.rs` file so
that it can be used in the end-to-end test.
* The `EvmAbi` is implemented. The Evm applications are untyped so there
is no point in adding specific types.
* The bytecode is written directly to file in order to have a system
that works in the same way as for Wasm.
* The `VmRuntime`is added to the `publish_and_create`.
* The code in `revm` is changed so that it suits the framework of Wasm:
- The `handle_query` is parsing the initial entry and returning the
parsed value.
- The `execute_operation` is not changed as it is put in the
`schedule_operation`.

The design of the parsing of json entries is not great. But it reflects
the design of json for the Wasm.

## Test Plan

The counter test using EVM is added to the `linera_net_tests.rs`.

## Release Plan

The functionality could be added to the TestNet.