Fix blob storage (#3502)
## Motivation

* As noted by @MathieuDutSik in #3496, LRU-caching is incorrect outside
views.
* However, journaling outside of views is also incorrect

## Proposal

On top of #3501, we simply deactivate the features that require
exclusive access to an object in storage.
* Journaling is not authorized.
* LRU cache should never insert `None` but instead forget about the
deleted key. (Not that we delete blobs but who knows in the future.)

We may want to rename `connect` and `clone_with_root_keys` later in
another PR.

## Test Plan

* CI
* Verified that this solves the issue with #3453

## Release Plan

In principle, this chould be backported to the latest `testnet` branch.

---------

Signed-off-by: Mathieu Baudet <1105398+ma2bd@users.noreply.github.com>
Co-authored-by: Andreas Fackler <afck@users.noreply.github.com>