Remove the need for `request_application` (#3382)
## Motivation

Until now, applications were only available on chains that published
them and ones that explicitly requested to use them using
`request_application`. This meant there was an awkward step for users
that wanted to use others' applications on their own chains.

## Proposal

This PR changes `UserApplicationDescription` into a blob. Publishing an
application means storing a blob, which is then accessible to all
microchains right away. This removes the need for `request_application`
and makes it possible to remove its functionality entirely.

## Test Plan

The example usage flow of the Request For Quotes application was tested
manually in order to ensure that the RFQ app was available on other
chains without it being explicitly requested.

The readme tests will ensure that this will be tested in CI as well.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- #2167
- #2362  
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)