Only allow HTTP requests authorized by the committee (#3427)
## Motivation

For security reasons, we would like to initially limit the hosts that
applications can send HTTP requests to.

## Proposal

Add an allow-list of hosts to the `ResourceControlPolicy` configured by
a committee. Only perform HTTP requests in the execution state actor if
the target host is in the list.

## Test Plan

Tests that exercise the allow list were added in PR #3509.

## Release Plan

- Nothing to do / These changes follow the usual release cycle, because
this is a new feature that should be included in the next release and
testnet.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)