Refactor how outgoing messages are exported by linera-execution (#3514)
## Motivation

Simplify the code some more after #3513

## Proposal

* Move the definition of `OutgoingMessage` to `linera-execution`
* Use `OutgoingMessage` instead of `ExecutionOutcome` in the transaction
tracker

`RawExecutionOutcome` is still use internally to `linera-execution` as a
way to create messages, notably in `system.rs` (and in tests). We should
probably also get rid of this concept at some point.

To be rebased on top of #3513

## Test Plan

CI

## Release Plan

This PR should not modify the protocol.