Fix epoch update in `TestValidator::change_resource_control_policy` (#3540)
## Motivation

Epoch changes were recently changed to use events (#3432), but the
`TestValidator::change_resource_control_policy` was still trying to
receive epoch change system messages instead of creating an operation to
process the new epoch.

## Proposal

Use the new `SystemOperation::ProcessNewEpoch` to update the chains to
use the new epoch, and skip updating the admin chain because it fails to
try to process the same epoch twice.

## Test Plan

Tested with a rebased version of the tests in #3509. The tests stop
working after #3432 but pass again after this PR.

## Release Plan

- These changes follow the usual release cycle, because it fixes an
issue with SDK that is not present in any released version.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)